refactor ast. extract display
