+++ src/ast/mod.rs
@@ -1,2 +1,3 @@
 pub mod program;
 pub mod parser;
+pub mod display;
